# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/documents": {
      "get": {
        "operationId": "DocumentsController_findAll",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Filtrer par catégorie",
            "schema": {
              "enum": [
                "FORMATIONS_PRO",
                "MIEUX_NOUS_CONNAITRE",
                "CGV"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des documents récupérée avec succès"
          }
        },
        "summary": "Récupérer tous les documents (accès public)",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/documents/{id}/download": {
      "get": {
        "operationId": "DocumentsController_download",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID du document",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document téléchargé avec succès"
          },
          "404": {
            "description": "Document non trouvé"
          }
        },
        "summary": "Télécharger un document (accès public)",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/documents/{id}/count": {
      "get": {
        "operationId": "DocumentsController_getDownloadCount",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID du document",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Nombre de téléchargements récupéré avec succès"
          },
          "404": {
            "description": "Document non trouvé"
          }
        },
        "summary": "Récupérer le nombre de téléchargements d'un document (accès public)",
        "tags": [
          "documents"
        ]
      },
      "post": {
        "operationId": "DocumentsController_incrementDownloadCount",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID du document",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Compteur incrémenté avec succès"
          },
          "404": {
            "description": "Document non trouvé"
          }
        },
        "summary": "Incrémenter le compteur de téléchargements (accès public)",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/documents/admin": {
      "post": {
        "operationId": "DocumentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "category": {
                    "type": "string",
                    "enum": [
                      "FORMATIONS_PRO",
                      "MIEUX_NOUS_CONNAITRE",
                      "CGV"
                    ]
                  },
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Document créé avec succès"
          },
          "400": {
            "description": "Données invalides ou fichier manquant"
          },
          "401": {
            "description": "Non autorisé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Créer un nouveau document (admin)",
        "tags": [
          "documents"
        ]
      },
      "get": {
        "operationId": "DocumentsController_findAllAdmin",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Filtrer par catégorie",
            "schema": {
              "enum": [
                "FORMATIONS_PRO",
                "MIEUX_NOUS_CONNAITRE",
                "CGV"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des documents récupérée avec succès"
          },
          "401": {
            "description": "Non autorisé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer tous les documents (admin)",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/documents/admin/{id}": {
      "get": {
        "operationId": "DocumentsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID du document",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document récupéré avec succès"
          },
          "401": {
            "description": "Non autorisé"
          },
          "404": {
            "description": "Document non trouvé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer un document par son ID (admin)",
        "tags": [
          "documents"
        ]
      },
      "put": {
        "operationId": "DocumentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID du document",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDocumentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Document mis à jour avec succès"
          },
          "401": {
            "description": "Non autorisé"
          },
          "404": {
            "description": "Document non trouvé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mettre à jour un document (admin)",
        "tags": [
          "documents"
        ]
      },
      "delete": {
        "operationId": "DocumentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID du document",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document supprimé avec succès"
          },
          "401": {
            "description": "Non autorisé"
          },
          "404": {
            "description": "Document non trouvé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Supprimer un document (admin)",
        "tags": [
          "documents"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Aloha Secourisme",
    "description": "\n      API REST complète pour la gestion de l'application Aloha Secourisme.\n      \n      Cette API permet de gérer les formations, les inscriptions, les documents, \n      la galerie d'images et l'administration du système.\n      \n      ## Authentification\n      \n      La plupart des endpoints nécessitent une authentification par token JWT.\n      Pour vous authentifier, utilisez l'endpoint `/api/auth/login` pour obtenir un token,\n      puis incluez ce token dans l'en-tête d'autorisation de vos requêtes.\n      \n      ## Rôles\n      \n      - **Accès public** : Accessible sans authentification\n      - **Admin** : Nécessite d'être connecté avec un compte administrateur\n      - **Super Admin** : Nécessite d'être connecté avec un compte super administrateur\n      \n      ## Modules principaux\n      \n      - **Formations** : Gestion des sessions de formation\n      - **Inscriptions** : Gestion des inscriptions aux formations\n      - **Documents** : Gestion des documents téléchargeables\n      - **Galerie** : Gestion des images de la galerie\n      - **News** : Gestion des actualités\n      - **Admin/Users** : Gestion des utilisateurs administrateurs\n      - **Dashboard** : Statistiques et tableaux de bord\n      - **Settings** : Paramètres du système\n      - **Database** : Export/import de la base de données\n    ",
    "version": "1.0.0",
    "contact": {
...
    },
...
    }
...
  "servers": [
...
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Entrez votre token JWT ici"
      }
    },
    "schemas": {
...
      "UpdateDocumentDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Fiche formation PSC1",
            "description": "Titre du document"
          },
          "category": {
            "type": "string",
            "enum": [
              "FORMATIONS_PRO",
              "MIEUX_NOUS_CONNAITRE",
              "CGV"
            ],
            "description": "Catégorie du document",
            "example": "FORMATIONS_PRO"
          }
        }
      },
...
    }
  }
}